extract logic from `Ruboclean::Runner` to `Ruboclean::ToYamlConverter`
