+++ lib/ruboclean.rb
@@ -2,8 +2,10 @@
 
 require "ruboclean/cli_arguments"
 require "ruboclean/orderer"
+require "ruboclean/grouper"
 require "ruboclean/path_cleanup"
 require "ruboclean/runner"
+require "ruboclean/to_yaml_converter"
 require "ruboclean/version"
 
 # Ruboclean entry point
