+++ lib/ruboclean/orderer.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "ruboclean/grouper"
-
 module Ruboclean
   # Orders the items within the groups alphabetically
   class Orderer
